+++ src/ui/menu/room/player_list_cell.hpp
@@ -1,32 +1,40 @@
 #pragma once
-#include <defs/all.hpp>
+
+#include <defs/geode.hpp>
 #include <data/types/gd.hpp>
 
-#include <ui/general/simple_player.hpp>
+class GlobedSimplePlayer;
 
 class PlayerListCell : public cocos2d::CCLayer {
 public:
     static constexpr float CELL_HEIGHT = 30.0f;
 
-    static PlayerListCell* create(const PlayerRoomPreviewAccountData& data, float width, bool forInviting);
+    static PlayerListCell* create(const PlayerRoomPreviewAccountData& data, float cellWidth, bool forInviting, bool isIconLazyLoad);
+
+    bool isIconLoaded();
+    void createPlayerIcon();
+    void createPlaceholderPlayerIcon();
 
 protected:
-    bool init(const PlayerRoomPreviewAccountData& data, float width, bool forInviting);
+    friend class RoomLayer;
+
+    bool init(const PlayerRoomPreviewAccountData& data, float cellWidth, bool forInviting, bool isIconLazyLoad);
     void onOpenProfile(cocos2d::CCObject*);
+
+    void sendInvite();
+    void enableInvites();
 
     void createInviteButton();
     void createJoinButton();
     void createAdminButton();
 
-    void sendInvite();
-    void enableInvites();
+    PlayerRoomPreviewAccountData playerData;
+    float cellWidth;
 
-    PlayerRoomPreviewAccountData data;
-    float width;
-
-    cocos2d::CCMenu* menu;
-    CCMenuItemSpriteExtra *playButton = nullptr, *inviteButton = nullptr;
-    GlobedSimplePlayer* simplePlayer;
-
-    bool isFriend;
+    cocos2d::CCMenu* rightButtonMenu;
+    cocos2d::CCMenu* leftSideLayout;
+    CCMenuItemSpriteExtra* joinButton = nullptr;
+    CCMenuItemSpriteExtra* inviteButton = nullptr;
+    GlobedSimplePlayer* simplePlayer = nullptr;
+    cocos2d::CCSprite* placeholderIcon = nullptr;
 };